% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{add_categorical_event}
\alias{add_categorical_event}
\title{Create Classification Variable Based On Window}
\usage{
add_categorical_event(varname = NULL, check_value = NULL, days = -365)
}
\arguments{
\item{varname}{variable name}

\item{check_value}{???}

\item{days}{size of window in days}
}
\value{
see above
}
\description{
takes categorical (string) data in events file and classifies person based on all available values in the window
before the person's index date (from cohort file).  Windows must be negative (i.e., before index date).  NA = FALSE
}
\examples{
\dontrun{# nothing here yet}
}

