% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{add_function_value}
\alias{add_function_value}
\title{Apply Function to Numeric Variable in Time Window}
\usage{
add_function_value(varname = NULL, days = -365, func = mean, ...)
}
\arguments{
\item{varname}{variable name}

\item{days}{size of window in days}

\item{func}{function to apply to values (e.g., mean)}

\item{...}{passed to function (e.g., na.rm = TRUE)}
}
\value{
see above
}
\description{
takes lab (or any numeric) data in events file and estimates function using all available values in the window
before the person's index date (from cohort file).  Windows must be negative (i.e., before index date).  Function
parameters can be passed in ... to the function (e.g., na.rm = TRUE).  Default function is mean
}
\examples{
\dontrun{# nothing here yet}
}

