% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{add_numerical_event}
\alias{add_numerical_event}
\title{Add Event to Cohort Using Function of Numeric Variable in Time Window}
\usage{
add_numerical_event(varname = NULL, days = -365, func = mean, ...)
}
\arguments{
\item{varname}{variable name}

\item{days}{size of window in days}

\item{func}{function to apply to values (e.g., mean)}

\item{...}{passed to function (e.g., na.rm = TRUE)}
}
\value{
see above
}
\description{
Takes any numeric data in events file (e.g., laboratory value) and estimates a function of it (e.g., mean)
using all available values.  Values can be limited to a time window before the person's index date (from cohort file).
Windows must be negative (i.e., before index date).  Function parameters can be passed in ... to the function
(e.g., na.rm = TRUE).  Default function is mean.  Other options include median, max, and min.
}
\examples{
\dontrun{# nothing here yet}
}
