% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{create_word_doc}
\alias{create_word_doc}
\title{Make Word File Using Template}
\usage{
create_word_doc(template_file = NULL)
}
\arguments{
\item{template_file}{File path to a .docx file that will be used as template.  Default is
to use the included template file.}
}
\value{
Returns a document that can be used with officer package
}
\description{
This simply uses the read_docx() function in officer package with the template file embedded
within this package.  An external template file can also be used.
}
\examples{
\dontrun{# nothing here yet}
}
