% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{fix_counts}
\alias{fix_counts}
\title{Fix Counts to Make 0}
\usage{
fix_counts(dt = cohort, dd = data_dict)
}
\arguments{
\item{dt}{data table}

\item{dd}{data dictionary file}
}
\value{
Corrected counts that have 0 for NA.  Calls change_NA
}
\description{
This function is designed to take a variable that reports NA for a count of 0 and replace
the NA with 0.  This occurs when Jigsaw finds no event (e.g., no hospitalization).  There
is now a flag variable reported in Jigsaw that indicates this explicitly.
}
\examples{
\dontrun{# nothing here yet}
}

