% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{fix_label_types}
\alias{fix_label_types}
\title{Fix Variable Types}
\usage{
fix_label_types(lab = labels)
}
\arguments{
\item{lab}{File with variable types and names}
}
\value{
Corrected label file for variable types
}
\description{
Convert label types to those in R.  note that fread can't read in dates and will use character instead
}
\examples{
\dontrun{# nothing here yet}
}

