% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{glancenames}
\alias{glancenames}
\title{Fix Tidyr "Glance" Names}
\usage{
glancenames(df)
}
\arguments{
\item{df}{Data frame from running tidyr glance()}
}
\value{
Data frame with proper column names
}
\description{
Fix Tidyr "Glance" Names
}
\examples{
\dontrun{# nothing here yet}
}

