% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_table1}
\alias{make_table1}
\title{Make "Table 1" to Summarize Variables in Data Frame}
\usage{
make_table1(varlist, dt)
}
\arguments{
\item{varlist}{List of variables to summarize}

\item{dt}{Data table with variables of interest}
}
\value{

}
\description{
Make "Table 1" to Summarize Variables in Data Frame
}
\examples{
\dontrun{# nothing here yet}
}

