% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{round_df}
\alias{round_df}
\title{Round All Numeric Columns in Data Frame}
\usage{
round_df(df, n = 3)
}
\arguments{
\item{df}{Data frame}

\item{n}{Digits to round to (to right of decimal)}
}
\value{
Data frame
}
\description{
Round All Numeric Columns in Data Frame
}
\examples{
\dontrun{# nothing here yet}
}

