% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{tidynames}
\alias{tidynames}
\title{Fix Tidyr "Tidy" Names}
\usage{
tidynames(df)
}
\arguments{
\item{df}{Data frame from running tidyr tidy()}
}
\value{
Data frame with proper column names
}
\description{
Fix Tidyr "Tidy" Names
}
\examples{
\dontrun{# nothing here yet}
}

