% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{wordtable}
\alias{wordtable}
\title{Create a Table in Word}
\usage{
wordtable(df, footer = NULL, header = NULL)
}
\arguments{
\item{df}{Data frame that will be the table}

\item{footer}{Footer text for the table}

\item{header}{Header text for the table}
}
\value{
Returns a FlexTable object.  To add this to a document, use the addtodoc() function
}
\description{
Create a Table in Word
}
\examples{
\dontrun{# nothing here yet}
}
