% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{addtodoc}
\alias{addtodoc}
\title{Add To Word Document}
\usage{
addtodoc(docu, item, caption = NULL, pagebreak = TRUE)
}
\arguments{
\item{docu}{Document object to which the object will be added}

\item{item}{Object from ReporteRs package to add}

\item{caption}{Text for the caption}

\item{pagebreak}{Should there be a pagebreak added (default is TRUE)}
}
\value{
Adds object to document
}
\description{
Adds a ReporteRs object to an existing document.  Generally, this will be a figure or a table, but it
could be anything supported by the ReporteRs package.
}
\examples{
\dontrun{# nothing here yet}
}

