% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{clean_numeric_event}
\alias{clean_numeric_event}
\title{Clean Numeric Event}
\usage{
clean_numeric_event(event_group = "", lower_cut = -Inf, upper_cut = Inf,
  unit_type = NULL)
}
\arguments{
\item{event_group}{event type to clean}

\item{lower_cut}{value below which data will be considered NA (inclusive)}

\item{upper_cut}{value above which data will be considered NA (inclusive)}

\item{unit_type}{limit to specific unit type (e.g., mmol/L)}
}
\value{
see above
}
\description{
Takes a numeric event type and makes all values <= lower cut or >= upper cut equal to "NA".  Also can
limit results to those with a specific unit type (which can be a character vector of multiple types)
Note -- need to add function to deduplicate by day and allow user to over-ride
}
\examples{
\dontrun{# nothing here yet}
}

