% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{fix_counts}
\alias{fix_counts}
\title{Fix Counts to Make 0}
\usage{
fix_counts(dt = cohort, dd = data_dict)
}
\arguments{
\item{dt}{data table}

\item{dd}{data dictionary file}
}
\value{
Corrected counts that have 0 for NA.  Calls change_NA
}
\description{
Fix Counts to Make 0
}
\examples{
\dontrun{# nothing here yet}
}

