% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{fmt}
\alias{fmt}
\title{Format Numeric Value as Character}
\usage{
fmt(x, digits = 2)
}
\arguments{
\item{x}{vector}

\item{digits}{digits to round}
}
\value{
The original numeric vector rounded, formatted and converted to character.
}
\description{
Format Numeric Value as Character
}
\examples{
\dontrun{# nothing here yet}
}
