% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{load_data}
\alias{load_data}
\title{Load Data from ZIP File}
\usage{
load_data(zipfile = NULL)
}
\arguments{
\item{zipfile}{The path to the ZIP file to load as a character string.}
}
\value{
Returns a list of the filenames for later loading.
}
\description{
This function takes a zipfile, identifies all files in the zipfile, and unzips them all into a temp file.
}
\examples{
\dontrun{# nothing here yet}
}

