% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{load_data}
\alias{load_data}
\title{Load Data from ZIP File}
\usage{
load_data(zipfile = NULL, load_files = TRUE, save_dir = ".", ver = 1L)
}
\arguments{
\item{zipfile}{The path to the ZIP file to load as a character string.}

\item{load_files}{Should the files be loaded to the global environment?  Defaults to TRUE.}

\item{save_dir}{The path to the directory where the file should be saved.  Defaults to working directory.}

\item{ver}{The version of the export being used. 1L is original.  2L is updated.}
}
\value{
Returns a list of the filenames for later loading.
}
\description{
This function takes a zipfile from Jigsaw, identifies all files in the zipfile, and unzips them all into a temp file.
Then it cleans the data by fixing dates to be an R date class (using the data.table IDate class for speed).  Finally,
it gives the option of loading all of the files into the global environment, or saving them all as RDS files
to a user-specified directory.  This function provides status updates at each step using the
R message function.  File reading is done using the data.table fread() function which is
set to "verbose" output to help the user identify problems.
}
\examples{
\dontrun{# nothing here yet}
}

