% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jigsaw.R
\name{make_dates}
\alias{make_dates}
\title{Converts Variable to Date Type}
\usage{
make_dates(dt = NULL, dd = data_dict)
}
\arguments{
\item{dt}{data table}

\item{dd}{The data dictionary with variable types.}
}
\value{
Variable conveted to an R date type
}
\description{
converts all date columns to date format (in place via data.table).  using data.table IDate class -- much faster
}
\examples{
\dontrun{# nothing here yet}
}

