% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_table1}
\alias{make_table1}
\title{Make "Table 1" to Summarize Variables in Data Frame}
\usage{
make_table1(varlist, dt, digits = 2)
}
\arguments{
\item{varlist}{Data frame containing 2 columns.  Column 1 is the names of the variables.
Column 2 is a name to be used as the variable name (typically longer than the variable name).}

\item{dt}{Data table with variables of interest in it.}

\item{digits}{Number of digits for continuous variables}
}
\value{
Returns Table 1 as a data.table
}
\description{
Make "Table 1" to Summarize Variables in Data Frame
}
\examples{
\dontrun{# nothing here yet}
}
