% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{round_df}
\alias{round_df}
\title{Round All Numeric Columns in Data Frame}
\usage{
round_df(df, n = 3)
}
\arguments{
\item{df}{Data frame}

\item{n}{Digits to round to (to right of decimal)}
}
\value{
Data frame
}
\description{
This was taken from this StackOverflow answer:  http://stackoverflow.com/questions/21328047/round-multiple-vectors-in-dataframe-with-plyr/21328269#21328269
}
\examples{
\dontrun{# nothing here yet}
}
