% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{round_if_numeric}
\alias{round_if_numeric}
\title{Round Numeric Variable}
\usage{
round_if_numeric(x, n = 3)
}
\arguments{
\item{x}{Vector to round}

\item{n}{Number of digits for rounding (default =)}
}
\value{
Vector with  numeric variable rounded
}
\description{
This was taken from this StackOverflow answer:  http://stackoverflow.com/questions/21328047/round-multiple-vectors-in-dataframe-with-plyr/21328269#21328269
}
\examples{
\dontrun{# nothing here yet}
}

